# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: NumCheck API
  description: |
    Event-driven, multi-tenant system for bulk validation of WhatsApp & Telegram numbers.
    
    ## Authentication
    
    The API supports two authentication methods:
    1. **JWT Token**: Use `Authorization: Bearer <token>` header
    2. **API Key**: Use `X-Api-Key: <api-key>` header
    
    ## Rate Limiting
    
    - Quick check: 60 requests/minute per tenant
    - Bulk operations: 200 items/second total processing
    - Rate limit headers are included in responses
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns error details in JSON format:
    ```json
    {
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: NumCheck Support
    email: support@numcheck.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.numcheck.com
    description: Production
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/v1/bulk/start:
    post:
      summary: Start bulk validation job
      description: |
        Start an asynchronous bulk validation job for up to 1 million numbers.
        Returns a job ID for tracking progress.
      tags: [Bulk Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkJobRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkJobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/bulk/status:
    get:
      summary: Get job status
      description: Get the current status and progress of a bulk validation job
      tags: [Bulk Operations]
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: string
          example: job_2025_0001
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/bulk/stream:
    get:
      summary: Stream job progress
      description: |
        Server-Sent Events stream for real-time job progress updates.
        Connect to this endpoint to receive live updates about job progress.
      tags: [Bulk Operations]
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: string
          example: job_2025_0001
      responses:
        '200':
          description: SSE stream of progress updates
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"jobId":"job_2025_0001","processed":1250,"total":10000,"progress":0.125}
                  
                  data: {"jobId":"job_2025_0001","processed":2500,"total":10000,"progress":0.25}
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/bulk/export.csv:
    get:
      summary: Export job results as CSV
      description: |
        Get a signed URL to download the job results as a CSV file.
        The URL is valid for 24 hours.
      tags: [Bulk Operations]
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: string
          example: job_2025_0001
      responses:
        '302':
          description: Redirect to signed S3 URL
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
...
  schemas:
...
    ValidationSummary:
      type: object
      properties:
        wa:
          type: object
          properties:
            registered:
              type: integer
            not_registered:
              type: integer
            business_active:
              type: integer
            unknown:
              type: integer
        tg:
          type: object
          properties:
            registered:
              type: integer
            not_registered:
              type: integer
            unknown:
              type: integer
    BulkJobRequest:
      type: object
      required: [numbers, platforms]
      properties:
        numbers:
          type: array
          items:
            type: string
          maxItems: 1000000
        platforms:
          type: array
          items:
            type: string
            enum: [whatsapp, telegram]
        countryDefault:
          type: string
          pattern: '^[A-Z]{2}$'
    BulkJobResponse:
      type: object
      properties:
        jobId:
          type: string
          example: "job_2025_0001"
    JobStatusResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, canceled]
        processed:
          type: integer
        total:
          type: integer
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 1
        summary:
          $ref: '#/components/schemas/ValidationSummary'
        exportUrl:
          type: string
          nullable: true
        duplicatesCount:
          type: integer
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              field: "numbers"
              message: "Array must contain at least 1 element"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            code: "UNAUTHORIZED"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when the rate limit resets (Unix timestamp)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded"
            code: "RATE_LIMITED"
            details:
              limit: 60
              remaining: 0
              resetTime: 1640995200
...